% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fits.R
\name{Coef_estim}
\alias{Coef_estim}
\title{Compute Coefficient Table with Standard Errors, Z-Values, and P-Values}
\usage{
Coef_estim(par, hessian = NULL, f_dist = NULL, y = NULL, name_par = NULL)
}
\arguments{
\item{par}{Numeric vector of parameter estimates.}

\item{hessian}{Optional numeric matrix. The Hessian evaluated at \code{par}.
If \code{NULL}, the Hessian is computed numerically from \code{f_dist} and \code{y}.}

\item{f_dist}{Optional function. Probability density function used to compute
the log-likelihood. \strong{The function must have arguments \code{theta} (parameters) and \code{x} (data).}}

\item{y}{Optional numeric vector. Observed data for computing the Hessian.}

\item{name_par}{Optional character vector. Names of the parameters.}
}
\value{
A data frame with columns:
\itemize{
\item Estimate: estimated parameter
\item Std. Error: standard error
\item z value: z-statistic
\item Pr(>|z|): p-value
\item Significance: significance codes
}
}
\description{
This function computes the standard errors, z-values, and p-values for a
set of parameter estimates given a Hessian matrix or a log-likelihood function.
}
\examples{
library(UMW)

par <- c(0.3, 1.3, 0.5)
name_par <- c("alpha", "gamma", "lambda")

# Example using the Hessian matrix:

hessian_ex <- -matrix(c(10, 2, 1, 2, 8, 0.5, 1, 0.5, 5),
                      nrow = 3, byrow = TRUE) # fictitious Hessian

Coef_estim(par = par, hessian = hessian_ex, name_par = name_par)

# Example using the probability density function:
\dontrun{

set.seed(1)
y <- runif(50, min = 0.001, max = 0.999)

dUMW1<-function(x,theta) {
  alpha<-theta[1]
  gamma<-theta[2]
  lambda<-theta[3]
  (alpha/(log(x)*x^(lambda+1)))*((-log(x))^gamma)*(lambda*log(x)-gamma)*
    exp(-alpha*((-log(x))^gamma)*x^(-lambda))
}

Coef_estim(par = par, f_dist = dUMW1, y = y, name_par = name_par)
}
}
