% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diagnostic.R
\name{diagn_DIST_RQUMW}
\alias{diagn_DIST_RQUMW}
\title{Distance-based influence diagnostics for RQ-UMW models}
\usage{
diagn_DIST_RQUMW(
  fit,
  M = c(1, 2),
  g_num = TRUE,
  base_size = 12,
  act_parallel = TRUE,
  influence = TRUE,
  print = TRUE,
  graphic = TRUE,
  n_cores = (parallel::detectCores() - 1),
  single_obs = NULL
)
}
\arguments{
\item{fit}{An object of class \code{RQUMW} returned by \code{fit_RQUMW()}.}

\item{M}{Integer vector indicating which distances to compute:
\code{1} for Frèchet distance and \code{2} for Rao distance.}

\item{g_num}{Logical; if \code{TRUE}, observation indices are displayed in plots.}

\item{base_size}{Base font size for diagnostic plots.}

\item{act_parallel}{Logical; if \code{TRUE}, parallel computation is used.}

\item{print}{Logical; if \code{TRUE}, results are printed to the console.}

\item{graphic}{Logical; if \code{TRUE}, diagnostic plots are produced.}

\item{n_cores}{Number of cores to be used in parallel computation.}
}
\value{
A list containing:
\itemize{
\item diagnostic measures (\eqn{Frèchet}, \eqn{Rao})
\item distance-based influence measures;
\item identified influential observations;
\item diagnostic plots (if requested).
}
}
\description{
Computes influence diagnostics based on distance measures between
fitted models with and without each observation. The implemented
distances are the Frèchet distance and the Rao distance.
}
\examples{
library(UMW)

set.seed(25)
X <- runif(100, 0, 1)
y <- 0.3 + 0.7*X + rnorm(100, 0, 0.05)
y <- pmin(pmax(y, 0.01), 0.99)
fit<-fit_RQUMW(f = y ~ X)

UMW::diagn_DIST_RUMW(fit,M = c(1,2))

UMW::diagn_DIST_RUMW(fit,M = c(2),g_num = F)

}
