% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GERA_UMW.R
\name{dUMW}
\alias{dUMW}
\title{Probability Density Function of the Unit-Modified Weibull(UMW) Distribution}
\usage{
dUMW(x, theta)
}
\arguments{
\item{x}{Numeric vector with values in the unit interval (0, 1).}

\item{theta}{Numeric vector of length 3 containing the parameter values
\eqn{(\alpha,\gamma,\lambda)}, all assumed to be positive.}
}
\value{
A numeric vector of the same length as \code{x} containing the values
of the probability density function evaluated at \code{x}.
}
\description{
Computes the probability density function (PDF) of the UMW distribution
for values in the unit interval.
}
\details{
The PDF is given by:

\deqn{
f(x;\alpha,\gamma,\lambda) =
\frac{\alpha}{\log(x)\,x^{\lambda+1}}
(-\log x)^{\gamma}
(\lambda \log x - \gamma)
\exp\left\{-\alpha (-\log x)^{\gamma} x^{-\lambda}\right\},
\quad 0 < x < 1,
}
where \eqn{\log(\cdot)} denotes the natural logarithm and
\eqn{\alpha}, \eqn{\gamma}, \eqn{\lambda > 0} are real-valued parameters.
}
\examples{
library(UMW)

x <- runif(n = 50,min = 0.001,max = 0.999)
theta <- c(alpha = 0.7, gamma = 1.3, lambda = 0.5)
dUMW(x, theta)

}
