% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GERA_UMW.R
\name{dUMW_order}
\alias{dUMW_order}
\title{Density of the r-th Order Statistic of the Unit-Modified Weibull(UMW) Distribution}
\usage{
dUMW_order(x, theta, r, n)
}
\arguments{
\item{x}{Numeric vector with values in the unit interval \eqn{(0,1)}.}

\item{theta}{Numeric vector of length 3 containing the parameters
\eqn{(\alpha,\gamma,\lambda)}.}

\item{r}{Integer indicating the order of the statistic
\eqn{(1 \le r \le n)}.}

\item{n}{Integer indicating the sample size.}
}
\value{
A numeric vector containing the values of the probability density
function of the \eqn{r}-th order statistic evaluated at \code{x}.
}
\description{
Computes the probability density function of the r-th order statistic
from a random sample of size \eqn{n} drawn from the UMW distribution.
}
\details{
The probability density function of the \eqn{r}-th order statistic is given by

\deqn{
f_{Y_{(r)}}(x) =
\frac{n!}{(r-1)!(n-r)!}
\frac{\alpha (-\log x)^{\gamma}}{\log(x)\, x^{\lambda+1}}
(\lambda \log x - \gamma)
\left[1 - \exp\{-\alpha (-\log x)^{\gamma} x^{-\lambda}\}\right]^{n-r}
\exp\left\{- r \alpha (-\log x)^{\gamma} x^{-\lambda}\right\},
\quad 0 < x < 1,
}

where \eqn{\log(\cdot)} denotes the natural logarithm and
\eqn{\alpha}, \eqn{\gamma}, \eqn{\lambda > 0}
are real-valued parameters.
}
\examples{
library(UMW)

theta <- c(alpha = 0.7, gamma = 1.3, lambda = 0.5)
n <- 10
r <- 3
x <- runif(n = 50,min = 0.001,max = 0.999)
dUMW_order(x, theta, r, n)

}
