% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GERA_UMW.R
\name{hUMW}
\alias{hUMW}
\title{Hazard Function of the Unit-Modified Weibull(UMW) Distribution}
\usage{
hUMW(x, theta)
}
\arguments{
\item{x}{Numeric vector with values in the unit interval \eqn{(0, 1)}.}

\item{theta}{Numeric vector of length 3 containing the parameters
\eqn{(\alpha,\gamma,\lambda)}.}
}
\value{
A numeric vector containing the values of the hazard function
evaluated at \code{x}.
}
\description{
Computes the hazard rate function of the UMW distribution.
}
\details{
The hazard function of the UMW distribution is defined as

\deqn{
h(x) =
\frac{
\alpha\, x^{-(\lambda+1)} (-\log x)^{\gamma}
(\lambda \log x - \gamma)
}{
\log(x)\left[\exp\left(\alpha (-\log x)^{\gamma} x^{-\lambda}\right)-1\right]
},
\quad 0 < x < 1,
}

where \eqn{\log(\cdot)} denotes the natural logarithm and
\eqn{\alpha}, \eqn{\gamma}, \eqn{\lambda > 0}
are real-valued parameters.
}
\examples{
library(UMW)

x <- runif(n = 50,min = 0.001,max = 0.999)
theta <- c(alpha = 0.7, gamma = 1.3, lambda = 0.5)
hUMW(x, theta)

}
