% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GERA_UMW.R
\name{pUMW}
\alias{pUMW}
\title{Cumulative Distribution Function of the Unit-Modified Weibull(UMW) Distribution}
\usage{
pUMW(x, theta)
}
\arguments{
\item{x}{Numeric vector with values in the unit interval \eqn{(0,1)}.}

\item{theta}{Numeric vector of length 3 containing the parameters
\eqn{(\alpha,\gamma,\lambda)}.}
}
\value{
A numeric vector containing the value of the cumulative distribution
function evaluated at each element of \code{x}.
}
\description{
Computes the cumulative distribution function (CDF) of the UMW distribution
for values in the unit interval.
}
\details{
The CDF is given by:

\deqn{
F(x;\alpha,\gamma,\lambda) =
\exp\left\{-\alpha (-\log x)^{\gamma} x^{-\lambda}\right\},
\quad 0 < x < 1,
}

where \eqn{\log(\cdot)} denotes the natural logarithm and
\eqn{\alpha}, \eqn{\gamma}, \eqn{\lambda > 0} are real-valued parameters.
}
\examples{
library(UMW)

x <- runif(n = 50,min = 0.001,max = 0.999)
theta <- c(alpha = 0.7, gamma = 1.3, lambda = 0.5)
pUMW(x, theta)

}
