% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gera_simu_RUMW_multi.R
\name{rRQUMW}
\alias{rRQUMW}
\title{Random generation from the reparameterized Unit-Modified Weibull distribution}
\usage{
rRQUMW(n, theta, X, Z, tau = 0.5, ginv_mu, ginv_lambda, set_seed = NULL)
}
\arguments{
\item{n}{Integer. Number of observations to be generated.}

\item{theta}{Numeric vector of model parameters.}

\item{X}{Matrix of covariates associated with the parameter \eqn{\mu}.}

\item{Z}{Matrix of covariates associated with the parameter \eqn{\lambda}.}

\item{tau}{Numeric. Quantile level used in the model. Default is 0.5.}

\item{ginv_mu}{Function. Inverse link function for \eqn{\mu}.}

\item{ginv_lambda}{Function. Inverse link function for \eqn{\lambda}.}

\item{set_seed}{Integer. Optional random seed for reproducibility.}
}
\value{
A numeric vector of length \code{n} containing random values generated
from the reparameterized UMW distribution.
}
\description{
Generates random observations from the reparameterized Unit Modified Weibull
distribution using numerical inversion of the cumulative distribution
function.
}
\examples{
set.seed(20)
n <- 50
X <- cbind(1,runif(n))
Z <- cbind(1,runif(n))
theta <- c(0.5,-0.7, 1.2, 0.4,2.8)

ginv_mu <- function(x) exp(x)/(1+exp(x))
ginv_lambda <- function(x) exp(x)

rRQUMW(n, theta, X, Z, tau = 0.5, ginv_mu, ginv_lambda,set_seed = 20)

}
